Added GET and POST methods to Newsletter API
Added a GET method to the Home resource that returns a welcome message.

Created a new Newsletters resource with GET and POST methods: - The GET method returns a list of all newsletters. - The POST method allows for the creation of new newsletters.

Created a new NewsletterByID resource with a GET method that returns a newsletter by its ID.